Move get_dashboard to dashboard/__init__
