+++ meinberlin/apps/dashboard2/__init__.py
@@ -1,9 +1,144 @@
+from functools import lru_cache
+
+from django.conf import settings
+from django.utils.module_loading import import_string
+
 from .components import DashboardComponent
 from .components import components
 from .components.forms import ModuleFormComponent
 from .components.forms import ModuleFormSetComponent
 from .components.forms import ProjectFormComponent
 
+
 __all__ = ['components', 'DashboardComponent',
            'ModuleFormComponent', 'ModuleFormSetComponent',
-           'ProjectFormComponent']
+           'ProjectFormComponent',
+           'get_project_dashboard']
+
+
+def get_project_dashboard(project):
+    key = 'PROJECT_DASHBOARD_CLASS'
+    dashboard_settings = getattr(settings, 'A4_DASHBOARD', None)
+    if dashboard_settings and key in dashboard_settings:
+        dashboard_cls = import_string(dashboard_settings[key])
+
+    else:
+        dashboard_cls = ProjectDashboard
+
+    return dashboard_cls(project)
+
+
+class ProjectDashboard:
+    def __init__(self, project):
+        self.project = project
+
+    def get_project_components(self):
+        return [component for component in components.get_project_components()
+                if component.get_menu_label(self.project)]
+
+    def get_module_components(self):
+        return components.get_module_components()
+
+    def get_progress(self):
+        num_valid = 0
+        num_required = 0
+
+        for component in self.get_project_components():
+            nums = component.get_progress(self.project)
+            num_valid = num_valid + nums[0]
+            num_required = num_required + nums[1]
+
+        for module in self.project.modules:
+            for component in self.get_module_components():
+                nums = component.get_progress(module)
+                num_valid = num_valid + nums[0]
+                num_required = num_required + nums[1]
+
+        return num_valid, num_required
+
+    def get_menu(self, current_module=None, current_component=None):
+        project_menu = self.get_project_menu(current_component)
+
+        menu_modules = []
+        for module in self.project.modules:
+            menu_module = self.get_module_menu(module, current_component,
+                                               current_module)
+            if menu_module:
+                menu_modules.append({
+                    'module': module,
+                    'menu': menu_module,
+                })
+
+        return {'project': project_menu, 'modules': menu_modules}
+
+    def get_project_menu(self, current_component):
+        project_menu = []
+        for component in self.get_project_components():
+            menu_item = component.get_menu_label(self.project)
+            if menu_item:
+                is_active = (component == current_component)
+                url = component.get_base_url(self.project)
+                num_valid, num_required = component.get_progress(self.project)
+                is_complete = (num_valid == num_required)
+
+                project_menu.append({
+                    'label': menu_item,
+                    'is_active': is_active,
+                    'url': url,
+                    'is_complete': is_complete,
+                })
+        return project_menu or None
+
+    def get_module_menu(self, module, current_module, current_component):
+        module_menu = []
+        for component in self.get_module_components():
+            menu_item = component.get_menu_label(module)
+            if menu_item:
+                is_active = (component == current_component and
+                             module.pk == current_module.pk)
+                url = component.get_base_url(module)
+                num_valid, num_required = component.get_progress(module)
+                is_complete = (num_valid == num_required)
+
+                module_menu.append({
+                    'label': menu_item,
+                    'is_active': is_active,
+                    'url': url,
+                    'is_complete': is_complete,
+                })
+        return module_menu or None
+
+
+# meinBerlin methods
+@lru_cache(maxsize=8)
+def get_project_type(project):
+    if (hasattr(project, 'externalproject') and
+            hasattr(project.externalproject, 'bplan')):
+        return 'bplan'
+    elif hasattr(project, 'externalproject'):
+        return 'external'
+    else:
+        return 'default'
+
+
+class TypedProjectDashboard(ProjectDashboard):
+    def __init__(self, project):
+        super().__init__(project)
+        self.project_type = get_project_type(project)
+
+    def get_project_components(self):
+        if self.project_type == 'bplan':
+            return [components.projects.get('bplan')]
+        elif self.project_type == 'external':
+            return [components.projects.get('external')]
+
+        return [component for component in components.get_project_components()
+                if component.get_menu_label(self.project)]
+
+    def get_module_components(self):
+        if self.project_type == 'bplan':
+            return []
+        elif self.project_type == 'external':
+            return []
+
+        return components.get_module_components()
